// packages/monaco-lsp-client/src/types.ts
export enum InlineCompletionTriggerKind {
  /**
   * Completion was triggered explicitly by a user gesture.
   */
  Invoked = 0,
  /**
   * Completion was triggered automatically while editing.
   */
  Automatic = 1,
}
// ...
export type LSPAny = LSPObject | LSPArray | string | number | boolean | null;
// ...
export type DocumentSelector = DocumentFilter[];
// ...
export interface WorkDoneProgressOptions {
  workDoneProgress?: boolean;
}

/**
 * General text document registration options.
 */
export interface TextDocumentRegistrationOptions {
  /**
   * A document selector to identify the scope of the registration. If set to null
   * the document selector provided on the client side will be used.
   */
  documentSelector: DocumentSelector | null;
}
// ...
export interface InlineCompletionItem {
  /**
   * The text to replace the range with. Must be set.
   */
  insertText: string | StringValue;
  /**
   * A text that is used to decide if this inline completion should be shown. When `falsy` the {@link InlineCompletionItem.insertText} is used.
   */
  filterText?: string;
  /**
   * The range to replace. Must begin and end on the same line.
   */
  range?: Range;
  /**
   * An optional {@link Command} that is executed *after* inserting this completion.
   */
  command?: Command;
}

/**
 * Inline completion options used during static or dynamic registration.
 *
 * @since 3.18.0
 * @proposed
 */
export interface InlineCompletionRegistrationOptions
  extends InlineCompletionOptions, TextDocumentRegistrationOptions, StaticRegistrationOptions {}
// ...
export interface Command {
  /**
   * Title of the command, like `save`.
   */
  title: string;
  /**
   * The identifier of the actual command handler.
   */
  command: string;
  /**
   * Arguments that the command handler should be
   * invoked with.
   */
  arguments?: LSPAny[];
}
// ...
export interface Range {
  /**
   * The range's start position.
   */
  start: Position;
  /**
   * The range's end position.
   */
  end: Position;
}

export type ImplementationOptions = WorkDoneProgressOptions;

/**
 * Static registration options to be returned in the initialize
 * request.
 */
export interface StaticRegistrationOptions {
  /**
   * The id used to register the request. The id can be used to deregister
   * the request again. See also Registration#id.
   */
  id?: string;
}
// ...
export interface Position {
  /**
   * Line position in a document (zero-based).
   *
   * If a line number is greater than the number of lines in a document, it defaults back to the number of lines in the document.
   * If a line number is negative, it defaults to 0.
   */
  line: number;
  /**
   * Character offset on a line in a document (zero-based).
   *
   * The meaning of this offset is determined by the negotiated
   * `PositionEncodingKind`.
   *
   * If the character value is greater than the line length it defaults back to the
   * line length.
   */
  character: number;
}
// ...
export interface StringValue {
  /**
   * The kind of string value.
   */
  kind: "snippet";
  /**
   * The snippet string.
   */
  value: string;
}

/**
 * Inline completion options used during static registration.
 *
 * @since 3.18.0
 * @proposed
 */
export type InlineCompletionOptions = WorkDoneProgressOptions;